+++ compute/acceptance/job_test.go
@@ -17,6 +17,73 @@ import (
 	"github.com/stretchr/testify/assert"
 )
 
+func TestAwsAccJobsCreate(t *testing.T) {
+	if _, ok := os.LookupEnv("CLOUD_ENV"); !ok {
+		t.Skip("Acceptance tests skipped unless env 'CLOUD_ENV' is set")
+	}
+
+	client := common.NewClientFromEnvironment()
+	jobsAPI := NewJobsAPI(context.Background(), client)
+
+	jobSettings := JobSettings{
+		NewCluster: &Cluster{
+			NumWorkers:   2,
+			SparkVersion: "6.4.x-scala2.11",
+			SparkConf:    nil,
+			AwsAttributes: &AwsAttributes{
+				Availability: "ON_DEMAND",
+			},
+			NodeTypeID: "r3.xlarge",
+		},
+		NotebookTask: &NotebookTask{
+			NotebookPath: "/Users/sri.tikkireddy@databricks.com/demo-terraform/demo-notebook",
+		},
+		Name: "1-sri-test-job",
+		Libraries: []Library{
+			{
+				Maven: &Maven{
+					Coordinates: "org.jsoup:jsoup:1.7.2",
+				},
+			},
+		},
+		EmailNotifications: &JobEmailNotifications{
+			OnStart:   []string{},
+			OnSuccess: []string{},
+			OnFailure: []string{},
+		},
+		TimeoutSeconds: 3600,
+		MaxRetries:     1,
+		Schedule: &CronSchedule{
+			QuartzCronExpression: "0 15 22 ? * *",
+			TimezoneID:           "America/Los_Angeles",
+		},
+		MaxConcurrentRuns: 1,
+	}
+
+	job, err := jobsAPI.Create(jobSettings)
+	assert.NoError(t, err, err)
+	id := job.ID()
+	defer func() {
+		err := jobsAPI.Delete(id)
+		assert.NoError(t, err, err)
+	}()
+	t.Log(id)
+	job, err = jobsAPI.Read(id)
+	assert.NoError(t, err, err)
+	assert.True(t, job.Settings.NewCluster.SparkVersion == "6.4.x-scala2.11",
+		"Something is wrong with spark version")
+
+	jobSettings.NewCluster.SparkVersion = "6.1.x-scala2.11"
+
+	err = jobsAPI.Update(id, jobSettings)
+	assert.NoError(t, err, err)
+
+	job, err = jobsAPI.Read(id)
+	assert.NoError(t, err, err)
+	assert.True(t, job.Settings.NewCluster.SparkVersion == "6.1.x-scala2.11",
+		"Something is wrong with spark version")
+}
+
 func TestAccJobResource(t *testing.T) {
 	if _, ok := os.LookupEnv("CLOUD_ENV"); !ok {
 		t.Skip("Acceptance tests skipped unless env 'CLOUD_ENV' is set")