+++ compute/resource_job.go
@@ -8,6 +8,7 @@ import (
 	"strings"
 
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
+	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/validation"
 
 	"github.com/databrickslabs/databricks-terraform/common"
 	"github.com/databrickslabs/databricks-terraform/internal"
@@ -96,6 +97,15 @@ var jobSchema = internal.StructToSchema(JobSettings{},
 			"`new_cluster` for greater reliability."
 		s["new_cluster"].Description = "Same set of parameters as for " +
 			"[databricks_cluster](cluster.md) resource."
+		p, err := internal.SchemaPath(s, "new_cluster", "num_workers")
+		if err == nil {
+			p.Optional = true
+			p.Default = 0
+			p.Type = schema.TypeInt
+			p.ValidateFunc = validation.IntAtLeast(0)
+			p.Required = false
+		}
+
 		if v, err := internal.SchemaPath(s, "new_cluster", "spark_conf"); err == nil {
 			v.DiffSuppressFunc = func(k, old, new string, d *schema.ResourceData) bool {
 				isPossiblyLegacyConfig := "new_cluster.0.spark_conf.%" == k && "1" == old && "0" == new
@@ -144,6 +154,12 @@ func ResourceJob() *schema.Resource {
 			if err != nil {
 				return err
 			}
+			if js.NewCluster != nil {
+				err = validateClusterDefinition(*js.NewCluster)
+				if err != nil {
+					return err
+				}
+			}
 			job, err := NewJobsAPI(ctx, c).Create(js)
 			if err != nil {
 				return err
@@ -163,6 +179,12 @@ func ResourceJob() *schema.Resource {
 			err := internal.DataToStructPointer(d, jobSchema, &js)
 			if err != nil {
 				return err
+			}
+			if js.NewCluster != nil {
+				err = validateClusterDefinition(*js.NewCluster)
+				if err != nil {
+					return err
+				}
 			}
 			return NewJobsAPI(ctx, c).Update(d.Id(), js)
 		},